collisions

+++ src/EngineObject.cpp
@@ -4,11 +4,11 @@ EngineObject::EngineObject() {
 	uid = getIDCounter()++;
 }
 
-unsigned long long& EngineObject::getIDCounter() {
-	static unsigned long long counter = 0;
+unsigned int& EngineObject::getIDCounter() {
+	static unsigned int counter = 0;
 	return counter;
 }
 
-unsigned long long EngineObject::getID() {
+unsigned int EngineObject::getID() {
 	return uid;
 }
